+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Changes
+
+* Update to Go 1.16
+
 ## 1.0.0 / 2021-04-23
 
 ### **Breaking Changes**